dto 추가, SecurityConfig 등등 수정

+++ src/main/java/com/example/demo/vo/Rq.java
@@ -49,7 +49,7 @@ public class Rq {
 
         this.loginedMember = member;
         System.out.println("rq");
-        this.loginedMemberId = (int) member.getId();
+        this.loginedMemberId = member.getId();
         this.loginedMemberNickName= member.getNickName();
         System.out.println("Nickname = " + loginedMember.getNickName());
         System.out.println("Nickname = " + loginedMemberNickName);
